connected both frontend and backend
