+++ client/src/App.jsx
@@ -19,4 +19,5 @@ function App() {
   );
 }
 
+
 export default App;
