+++ server/ASAP user/post.bru
@@ -1,0 +1,23 @@
+meta {
+  name: post
+  type: http
+  seq: 2
+}
+
+put {
+  url: http://localhost:5000/api/users/6811c4b9f647554016dbe38c
+  body: json
+  auth: inherit
+}
+
+body:json {
+  
+    {
+      "username": "sandanam k",
+      "email": "sand123y@gmail.com",
+      "password": "1234567",
+      "role": "user"
+    }
+  
+  
+}
