+++ server/models/postschema.js
@@ -1,10 +1,11 @@
 const mongoose = require('mongoose');
+
 const UserSchema = new mongoose.Schema({
   username: { type: String, required: true, unique: true },
   email: { type: String, required: true, unique: true },
   password: { type: String, required: true },
-  role: { type: String, enum: ['user', 'admin'], default: 'user' }, 
-  createdAt: { type: Date, default: Date.now },
+  role: { type: String, enum: ['user', 'admin'], default: 'user' },
+  
 });
 
 module.exports = mongoose.model('User', UserSchema);
