+++ server/package.json
@@ -11,6 +11,10 @@
   "author": "Sandanam",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.21.2"
+    "cors": "^2.8.5",
+    "dotenv": "^16.5.0",
+    "express": "^4.21.2",
+    "mogoose": "^0.0.1-security",
+    "mongoose": "^8.14.0"
   }
 }
