+++ server/server.js
@@ -1,0 +1,32 @@
+require('dotenv').config();
+const express = require('express');
+const cors = require('cors');
+const mongoose = require('mongoose');
+const userRoutes = require('./routes/routes');
+
+const app = express();
+const port = process.env.PORT || 5000;
+
+// Middlewares
+app.use(cors());
+app.use(express.json());
+
+// DB connection
+mongoose.connect(process.env.MONGO_URI, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true
+})
+  .then(() => console.log('✅ MongoDB connected'))
+  .catch((err) => console.error('❌ MongoDB connection error:', err));
+
+// Routes
+app.get('/ping', (req, res) => {
+  res.send('Pong!');
+});
+
+app.use('/api', userRoutes);
+
+// Start server
+app.listen(port, () => {
+  console.log(`🚀 Server is running at http://localhost:${port}`);
+});
